+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - improvement: Allow the admin to update payment method only the overdue subscription item without cancel PayZen subscription
+- improvement: add payment transfer/check to accounting settings
 
 ## v6.3.11 2024 February 2